Switch rsign download to different place
